+++ R/pdb.R
@@ -6,7 +6,7 @@
 #' \sQuote{pdb}. The purpose of this class is to store the data of molecular
 #' systems contained in PDB files. The default method of the \code{pdb} function
 #' creates an object of class \sQuote{pdb} from its different components, i.e.:
-#' \code{title}, \code{remark}, \code{cryst1}, \code{atoms} and \code{conect}. 
+#' \code{title}, \code{remark}, \code{crystal}, \code{atoms} and \code{conect}. 
 #' At least an object of class \sQuote{atoms} has to be specified. \cr\cr 
 #' \code{is.pdb} tests if x is an object of class \sQuote{pdb}, i.e. if x has a
 #' \dQuote{class} attribute equal to \code{pdb}.
@@ -15,14 +15,14 @@
 #' \code{pdb} returns a list of class \sQuote{pdb} with the following components:
 #' \item{title}{a character vector containing the TITLE records found in a PDB file.}
 #' \item{remark}{a character vector containing the REMARK records found in a PDB file.}
-#' \item{cryst1}{a list of class \sQuote{cryst1} containing the first CRYST1 record found in a PDB file. All others are ignored.}
+#' \item{crystal}{a list of class \sQuote{crystal} containing the first CRYST1 record found in a PDB file. All others are ignored.}
 #' \item{atoms}{a data.frame of class \sQuote{atoms} containing the ATOM and HETATM records found in a PDB file.}
 #' \item{conect}{a data.frame of class \sQuote{conect} containing the CONECT records found in a PDB file.}
 #' \cr
 #' \code{is.pdb} returns TRUE if x is an object of class \sQuote{pdb} and FALSE otherwise.
 #' 
 #' @param atoms a data.frame of class \code{atoms} containing ATOM and HETATM records use to create the \code{pdb} object.
-#' @param crystal a list of class \code{cryst1} containing the periodical boundary conditions and space group used to create the \code{pdb} object.
+#' @param crystal a list of class \code{crystal} containing the periodical boundary conditions and space group used to create the \code{pdb} object.
 #' @param conect a data.frame of class \code{conect} containing CONECT records use to create the \code{pdb} object.
 #' @param remark a character vector containing some REMARK records to be added to the \code{pdb} object.
 #' @param title a character vector containing some TITLE records to be added to the \code{pdb} object.
@@ -32,12 +32,12 @@
 #' @param cryst1 will be deprecated and replaced by argument crystal.
 #' 
 #' @seealso 
-#' \code{\link{atoms}}, \code{\link{coords}}, \code{\link{cryst1}}, \code{\link{conect}} and \code{\link{read.pdb}}
+#' \code{\link{atoms}}, \code{\link{coords}}, \code{\link{crystal}}, \code{\link{conect}} and \code{\link{read.pdb}}
 #' 
 #' @examples 
 #' title  <- "This is just an example"
 #' remark <- NULL
-#' cryst1 <- cryst1(c(10,10,10))
+#' cryst1 <- crystal(c(10,10,10))
 #' atoms <- atoms(recname = c("ATOM","ATOM"), eleid = 1:2, elename = c("H","H"), alt = "",
 #'                resname = c("H2","H2"), chainid = "", resid = c(1,1), insert = "",
 #'                x1 = c(0,0), x2 = c(0,0), x3 = c(0,1), occ = c(0.0,0.0), temp = c(1.0,1.0),
@@ -62,14 +62,7 @@ pdb.default <- function(atoms, crystal = NULL, conect = NULL, remark = NULL, tit
 	if( ! is.atoms(atoms)) stop("'atoms' must be an object of class 'atoms'")
 	
 	# Crystal cell:
-	if( ! is.null(cryst1)) {
-		if(is.null(crystal)) {
-			crystal = cryst1;
-		} else {
-			stop("Please specify only crystal!",
-				"Note: arg cryst1 will be deprecated!");
-		}
-	}
+	crystal = checkArgCrystal(crystal, cryst1);
 	if( ! is.null(crystal) & ! is.crystal(crystal))
 		stop("'crystal' must be an object of class 'crystal'");
 	if( ! is.null(conect) & ! is.conect(conect))
@@ -81,11 +74,12 @@ pdb.default <- function(atoms, crystal = NULL, conect = NULL, remark = NULL, tit
   if(! is.character(title ) & ! is.null(title )) title  <- as.character(title )
   if(! is.character(remark) & ! is.null(remark)) remark <- as.character(remark)
   
-  to.return <- list(title = title, remark = remark, cryst1 = crystal, atoms = atoms, conect = conect)
-  if( ! is.null(resolution)) to.return$Resolution = resolution;
-  
-  class(to.return) <- c("pdb", "list")
-  return(to.return)
+	to.return = list(title = title, remark = remark,
+		crystal = crystal, atoms = atoms, conect = conect);
+	if( ! is.null(resolution)) to.return$Resolution = resolution;
+	
+	class(to.return) = c("pdb", "list");
+	return(to.return);
 }
 
 #' @rdname pdb