+++ R/read.pdb.R
@@ -17,7 +17,7 @@
 #' When a single MODEL section is read, this function returns an object of class  \sQuote{pdb} (a list with a \code{class} attribute equal to \code{pdb}) with the following components:
 #' \item{title}{a character vector containing the TITLE records found in the PDB file.}
 #' \item{remark}{a character vector containing the REMARK records found in the PDB file.}
-#' \item{cryst1}{a list of class \sQuote{cryst1} containing the first CRYST1 record found in the PDB file. All others are ignored.}
+#' \item{crystal}{a list of class \sQuote{crystal} containing the first CRYST1 record found in the PDB file. All others are ignored.}
 #' \item{atoms}{a data.frame of class \sQuote{atoms} containing the ATOM and HETATM records found in the PDB file.}
 #' \item{conect}{a data.frame of class \sQuote{conect} containing the CONECT records found in the PDB file.}
 #' When multiple MODEL sections are read, a list of object of class \sQuote{pdb} is returned.
@@ -39,7 +39,7 @@
 #' http://www.wwpdb.org/documentation/format33/v3.3.html
 #' 
 #' @seealso
-#' \code{\link{write.pdb}}, \code{\link{pdb}}, \code{\link{cryst1}}, \code{\link{atoms}}, \code{\link{conect}}
+#' \code{\link{write.pdb}}, \code{\link{pdb}}, \code{\link{crystal}}, \code{\link{atoms}}, \code{\link{conect}}
 #' 
 #' @examples 
 #' # Read a PDB file included with the package
@@ -72,10 +72,9 @@ read.pdb <- function(file, ATOM = TRUE, HETATM = TRUE, CRYSTAL = TRUE,
 		}
 	}
 	if(!file.exists(file))
-		stop("File '", file, "'' is missing");
+		stop("File '", file, "' is missing!");
 	
 	lines = readLines(file);
-	
 	recname = substr(lines, 1, 6);
 	
 	trim = function(str)