+++ R/rotation.R
@@ -3,9 +3,9 @@
 #' Rotation of atomic coordinates around a given vector.
 #' 
 #' \code{R} is generic functions. Method for objects of class \sQuote{coords}
-#' first convert the coordinates into Cartesian coordinates using \code{cryst1}
+#' first convert the coordinates into Cartesian coordinates using \code{crystal}
 #' if needed. Once rotated, the coordinates are reconverted back to the orginal
-#' basis set using again \code{cryst1}. Method for objects of class \sQuote{pdb}
+#' basis set using again \code{crystal}. Method for objects of class \sQuote{pdb}
 #' first extract coordinates from the object using the function \code{coords},
 #' perform the rotation, and update the coordinates of the \sQuote{pdb} object
 #' using the function \code{coords<-}.
@@ -18,7 +18,7 @@
 #' @param y the y-component of the rotation vector.
 #' @param z the z-component of the rotation vector.
 #' @param mask a logical vector indicating the set of coordinates to which the rotation has to be applyed.
-#' @param cryst1 an object of class \sQuote{cryst1} use to convert fractional into Cartesian coordinates when need.
+#' @param cryst1 an object of class \sQuote{crystal} used to convert fractional into Cartesian coordinates (when needed).
 #' @param \dots further arguments passed to or from other methods.
 #' 
 #' @seealso 
@@ -73,7 +73,8 @@ R.coords <- function(obj, angle = 0, x = 0, y = 0, z = 1, mask = TRUE, cryst1 = 
 
 #' @rdname rotation
 #' @export
-R.pdb <- function(obj, angle = 0, x = 0, y = 0, z = 1, mask = TRUE, cryst1 = obj$cryst1, ...){
+R.pdb <- function(obj, angle = 0, x = 0, y = 0, z = 1, mask = TRUE,
+		cryst1 = obj$crystal, ...) {
   if(!is.pdb(obj)) stop("'object' must be an obj of class 'pdb'")
   
   coords(obj) <- R(coords(obj), angle = angle, x = x, y = y, z = z, mask = mask, cryst1 = cryst1, ...)